Move shareLink function
